Unit Converter
I made a unit converter, which can convert some often used variables. In the making of this project I learned more about for and while loops. I used while loops, to make the program rerun and while loops, to validate user input. I used the for loops to iterate the tuples, which defines the conversions. Afterwards a new for loop prints the options as an f-string, so that I can print the variables as str automatically.